+++ src/stream/random.rs
@@ -44,8 +44,16 @@ impl InnerStreamCipherAlgorithm {
 pub fn default_stream_cipher() -> (impl StreamCipher, Vec<u8>) {
     let mut key = vec![0u8; 64];
     OsRng.fill_bytes(&mut key);
-    let iv = Sha256::digest(key.as_ref());
+    let iv = Sha512::digest(key.as_ref());
     let cipher = ChaCha20::new_var(&iv[0..32], &iv[32..44]).unwrap();
 
     (cipher, key)
+}
+
+/// Create a default stream cipher with a given key
+///
+/// The stream cipher is created using ChaCha20, and the key is generated from OS randomness.
+pub fn default_stream_cipher_with_key(key: Vec<u8>) -> impl StreamCipher {
+    let iv = Sha512::digest(key.as_ref());
+    ChaCha20::new_var(&iv[0..32], &iv[32..44]).unwrap()
 }