+++ src/xml.rs
@@ -4,6 +4,6 @@ mod decoders;
 pub(crate) mod parse;
 pub(crate) mod serialize;
 
-pub use crate::stream::random::{default_stream_cipher, InnerStreamError};
+pub use crate::stream::random::{default_stream_cipher, default_stream_cipher_with_key, InnerStreamError};
 pub use parse::parse_xml;
 pub use serialize::write_xml;