+++ src/xml/serialize.rs
@@ -2,6 +2,7 @@ use super::decoders::{encode_datetime, encode_uuid};
 use crate::types::{Database, Entry, Field, Group, MemoryProtection, Meta, Times, Value};
 use std::io::Write;
 use thiserror::Error;
+use stream_cipher::StreamCipher;
 use xml::writer::events::XmlEvent;
 use xml::writer::EventWriter as XmlWriter;
 
@@ -33,13 +34,16 @@ fn write_string_tag<W: Write, S: AsRef<str>>(
     Ok(())
 }
 
-fn write_field<W: Write>(writer: &mut XmlWriter<W>, wrapper: &str, field: &Field) -> Result<()> {
+fn write_field<W: Write, S: StreamCipher + ?Sized>(writer: &mut XmlWriter<W>, wrapper: &str, field: &Field, stream_cipher: &mut S) -> Result<()> {
     writer.write(XmlEvent::start_element(wrapper))?;
     write_string_tag(writer, "Key", &field.key)?;
     match &field.value {
         Value::Protected(v) => {
             writer.write(XmlEvent::start_element("Value").attr("Protected", "True"))?;
-            writer.write(XmlEvent::characters(&v))?;
+            let mut encrypt_buf = v.clone().into_bytes();
+            stream_cipher.encrypt(&mut encrypt_buf);
+            let encrypted = base64::encode(encrypt_buf);
+            writer.write(XmlEvent::characters(&encrypted))?;
             writer.write(XmlEvent::end_element())?;
         }
         Value::Standard(v) => write_string_tag(writer, "Value", &v)?,
@@ -66,14 +70,14 @@ fn write_memory_protection<W: Write>(
     Ok(())
 }
 
-fn write_meta<W: Write>(writer: &mut XmlWriter<W>, meta: &Meta) -> Result<()> {
+fn write_meta<W: Write, S: StreamCipher + ?Sized>(writer: &mut XmlWriter<W>, meta: &Meta, stream_cipher: &mut S) -> Result<()> {
     writer.write(XmlEvent::start_element("Meta"))?;
     write_string_tag(writer, "Generator", "kdbx-rs")?;
     write_string_tag(writer, "DatabaseName", &meta.database_name)?;
     write_string_tag(writer, "DatabaseDescription", &meta.database_description)?;
     writer.write(XmlEvent::start_element("CustomData"))?;
     for field in &meta.custom_data {
-        write_field(writer, "Item", field)?;
+        write_field(writer, "Item", field, stream_cipher)?;
     }
     writer.write(XmlEvent::end_element())?;
     write_memory_protection(writer, &meta.memory_protection)?;
@@ -106,17 +110,17 @@ fn write_times<W: Write>(writer: &mut XmlWriter<W>, times: &Times) -> Result<()>
     Ok(())
 }
 
-fn write_entry<W: Write>(writer: &mut XmlWriter<W>, entry: &Entry) -> Result<()> {
+fn write_entry<W: Write, S: StreamCipher + ?Sized>(writer: &mut XmlWriter<W>, entry: &Entry, stream_cipher: &mut S) -> Result<()> {
     writer.write(XmlEvent::start_element("Entry"))?;
     write_string_tag(writer, "UUID", &encode_uuid(&entry.uuid))?;
     write_times(writer, &entry.times)?;
     for field in &entry.fields {
-        write_field(writer, "String", field)?;
+        write_field(writer, "String", field, stream_cipher)?;
     }
     if !entry.history.is_empty() {
         writer.write(XmlEvent::start_element("History"))?;
         for old_entry in &entry.history {
-            write_entry(writer, old_entry)?;
+            write_entry(writer, old_entry, stream_cipher)?;
         }
         writer.write(XmlEvent::end_element())?;
     }
@@ -124,32 +128,32 @@ fn write_entry<W: Write>(writer: &mut XmlWriter<W>, entry: &Entry) -> Result<()>
     Ok(())
 }
 
-fn write_group<W: Write>(writer: &mut XmlWriter<W>, group: &Group) -> Result<()> {
+fn write_group<W: Write, S: StreamCipher + ?Sized>(writer: &mut XmlWriter<W>, group: &Group, stream_cipher: &mut S) -> Result<()> {
     writer.write(XmlEvent::start_element("Group"))?;
     write_string_tag(writer, "UUID", encode_uuid(&group.uuid))?;
     write_string_tag(writer, "Name", &group.name)?;
     write_times(writer, &group.times)?;
     for entry in &group.entries {
-        write_entry(writer, &entry)?;
+        write_entry(writer, &entry, stream_cipher)?;
     }
     for group in &group.children {
-        write_group(writer, &group)?;
+        write_group(writer, &group, stream_cipher)?;
     }
     writer.write(XmlEvent::end_element())?;
     Ok(())
 }
 
 /// Write the decrypted XML for a database to a file
-pub fn write_xml<W: Write>(output: W, database: &Database) -> Result<()> {
+pub fn write_xml<W: Write, S: StreamCipher + ?Sized>(output: W, database: &Database, stream_cipher: &mut S) -> Result<()> {
     let config = xml::EmitterConfig::default()
         .perform_indent(true)
         .indent_string("\t");
     let mut writer = xml::EventWriter::new_with_config(output, config);
     writer.write(XmlEvent::start_element("KeePassFile"))?;
-    write_meta(&mut writer, &database.meta)?;
+    write_meta(&mut writer, &database.meta, stream_cipher)?;
     writer.write(XmlEvent::start_element("Root"))?;
     for group in &database.groups {
-        write_group(&mut writer, group)?;
+        write_group(&mut writer, group, stream_cipher)?;
     }
     writer.write(XmlEvent::end_element())?;
     writer.write(XmlEvent::end_element())?;