+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.2.0
+
+* Add support for writing protected values
+* New APIs;
+  * `Group::recursive_entries()` / `Group::recursive_entries_mut()` for iterating
+    through all entries in a group and its child groups
+
 ## 0.1.2
 
 * Add support for reading KeepassXC OTP secrets (no code generation yet)